refactor: 페이지 컴포넌트 분리 및 공통 캐러셀 컴포넌트 추출
