+++ src/routes/tester/test/$testId/index.tsx
@@ -1,7 +1,7 @@
-import { useRef, useState } from "react";
 import { createFileRoute } from "@tanstack/react-router";
-import { Asset, BottomCTA, ListRow, Paragraph, Top } from "@toss/tds-mobile";
+import { BottomCTA, ListRow, Paragraph, Top } from "@toss/tds-mobile";
 import { adaptive } from "@toss/tds-colors";
+import { ImageCarousel } from "@/shared/components/ImageCarousel";
 
 const DUMMY_IMAGES = [
   "https://static.toss.im/3d-emojis/u1F35C.png",
@@ -15,26 +15,10 @@ export const Route = createFileRoute("/tester/test/$testId/")({
 
 function TestDetailPage() {
   // const { testId } = Route.useParams();
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const touchStartX = useRef(0);
-
-  const handleTouchStart = (e: React.TouchEvent) => {
-    touchStartX.current = e.touches[0].clientX;
-  };
-
-  const handleTouchEnd = (e: React.TouchEvent) => {
-    const diff = touchStartX.current - e.changedTouches[0].clientX;
-    if (Math.abs(diff) < 30) return;
-    if (diff > 0) {
-      setCurrentIndex((i) => Math.min(i + 1, DUMMY_IMAGES.length - 1));
-    } else {
-      setCurrentIndex((i) => Math.max(i - 1, 0));
-    }
-  };
 
   return (
     <div className="flex flex-col min-h-screen">
-      <div className="flex-1 overflow-y-auto pb-[90px]">
+      <div className="flex-1 overflow-y-auto pb-22.5">
         <Top
           title={
             <Top.TitleParagraph size={22} color={adaptive.grey900}>
@@ -61,40 +45,8 @@ function TestDetailPage() {
           }
         />
 
-        {/* 썸네일 캐러셀 */}
-        <div className="flex flex-col items-center px-4 gap-2">
-          <div
-            className="w-full h-[193px] rounded-2xl bg-cover bg-center"
-            style={{
-              backgroundImage: `url(${DUMMY_IMAGES[currentIndex]})`,
-              boxShadow:
-                "inset 0 0 0 1px var(--token-tds-color-grey-opacity-100, rgba(2,32,71,0.05))",
-            }}
-            onTouchStart={handleTouchStart}
-            onTouchEnd={handleTouchEnd}
-          />
+        <ImageCarousel images={DUMMY_IMAGES} />
 
-          {/* 페이지 인디케이터 */}
-          <div className="flex flex-row gap-1 items-center p-2 mb-3">
-            {DUMMY_IMAGES.map((_, i) => (
-              <Asset.Icon
-                key={i}
-                frameShape={{ width: 12, height: 12 }}
-                backgroundColor="transparent"
-                name="icon-circle-16-mono"
-                color={
-                  i === currentIndex
-                    ? adaptive.greyOpacity500
-                    : adaptive.greyOpacity300
-                }
-                aria-hidden={true}
-                ratio="1/1"
-              />
-            ))}
-          </div>
-        </div>
-
-        {/* 보상 머니 */}
         <ListRow
           left={
             <ListRow.AssetIcon
@@ -116,7 +68,6 @@ function TestDetailPage() {
           horizontalPadding="small"
         />
 
-        {/* 한 줄 소개 */}
         <ListRow
           left={
             <ListRow.AssetIcon name="icon-document-line-fill" variant="fill" />
